Part 1: Cleanup styles & stuff
